requirements.txt
absl-py==2.1.0
accelerate==0.29.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
black==24.3.0
Brotli
certifi
charset-normalizer
click==8.1.7
clip==1.0
cloudpickle==3.0.0
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
cylp==0.92.2
Cython==3.0.10
fonttools==4.51.0
ftfy==6.2.0
fvcore==0.1.5.post20221221
grpcio==1.62.1
h5py==3.10.0
huggingface-hub
humanfriendly==10.0
hydra-core==1.3.2
idna
imageio==2.34.0
immutabledict==4.2.0
importlib-metadata
iopath==0.1.9
Jinja2
joblib
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite
Markdown==3.6
MarkupSafe
matplotlib==3.8.4
mkl-fft 
mkl-random 
mkl-service==2.4.0
mpmath
MultiScaleDeformableAttention==1.0
munkres==1.1.4
mypy-extensions==1.0.0
networkx
numba 
numpy 
omegaconf==2.3.0
opencv-python==4.9.0.80
ortools==9.3.10497
packaging
pandas==2.2.1
pathspec==0.12.1
pillow
platformdirs==4.2.0
portalocker==2.8.2
POT==0.9.3
protobuf==5.26.1
psutil==5.9.8
pycocotools==2.0.7
pyparsing==3.1.2
pyquaternion==0.9.9
PySocks
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML
regex 
requests
safetensors
scikit-image==0.22.0
scikit-learn
scipy 
shapely==2.0.3
six==1.16.0
submitit==1.5.1
sympy 
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl
tifffile==2024.2.12
timm==0.9.16
tokenizers 
tomli==2.0.1
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tqdm 
transformers
typing==3.7.4.3
typing_extensions
tzdata==2024.1
urllib3 
wcwidth==0.2.13
Werkzeug==3.0.2
yacs==0.1.8
zipp 
